requirements.txt
PyMySQL==1.0.2
PyYAML==6.0.1
PyYAML==6.0.1
redis==5.0.2
selenium==4.8.2

